Allow item action priority to be long instead of int

+++ _next/Fluffle.Ingestion.Core/Domain/ItemActions/ItemAction.cs
@@ -6,7 +6,7 @@ public abstract class ItemAction
 
     public required string ItemId { get; set; }
 
-    public required int Priority { get; set; }
+    public required long Priority { get; set; }
 
     public required int AttemptCount { get; set; }
